+++ src/mbapp.cpp
@@ -1,4 +1,4 @@
-#include "mbapp.h"
+#include "header/mbapp.h"
 
 #ifdef _WIN32
 #include <Windows.h>
